requirements.txt
beautifulsoup4>=4.12,<5.0
faiss-cpu>=1.8,<2.0
fastapi>=0.111,<1.0
openai>=1.30,<2.0
pydantic>=2.7,<3.0
PyMuPDF>=1.24,<2.0
python-dotenv>=1.0,<2.0
pytest>=8.0,<9.0
requests>=2.32,<3.0
streamlit>=1.36,<2.0
uvicorn>=0.30,<1.0
